+++ packages/wallet/src/initialization/instances/index.ts
@@ -7,3 +7,4 @@ export { loggingController } from './logging-controller/logging-controller';
 export { networkController } from './network-controller/network-controller';
 export { remoteFeatureFlagController } from './remote-feature-flag-controller/remote-feature-flag-controller';
 export { storageService } from './storage-service/storage-service';
+export { transactionController } from './transaction-controller/transaction-controller';